telemetry: attempts at fixes
The primary change here is to increase the rx buffer size, to ensure
that if we are bound up trying to send an object we can at least receive
one other frame of requests before we start dropping client request
info.

Also make GCS retry one more time, and on F3 deal with short-reports
better than we previously did.